requirements.txt
# Core dependencies
pocketflow>=0.0.1
gradio>=4.0.0

# LLM providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Data processing
PyYAML>=6.0
lxml>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# NLP and text processing
spacy>=3.7.0
nltk>=3.8

# Performance monitoring and system metrics
psutil>=5.9.0

# Web and async
websockets>=11.0.0
asyncio

# Data analysis and visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0

# Standard library modules (no version needed)
# hashlib, json, datetime, typing, re, os, functools, logging